KeyError: 'timestampMS' Bug Fixed
